0.0.1

+++ graphs/graph_time.py
@@ -18,5 +18,5 @@ class graph_time(pg.PlotItem):
 
     def update(self, value):
         self.time_text.setText('')
-        self.tiempo = round(abs(int(value)) / 1000, 2)
+        self.tiempo = round(abs(float(value)) / 1000, 2)
         self.time_text.setText(str(self.tiempo))
